+++ packages/article/src/article-body/article-body-row.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Dimensions, StyleSheet, View } from "react-native";
+import { Dimensions, View } from "react-native";
 import PropTypes from "prop-types";
 import ArticleImage from "@times-components/article-image";
 import Ad from "@times-components/ad";
@@ -8,33 +8,11 @@ import KeyFacts from "@times-components/key-facts";
 import { renderTree } from "@times-components/markup-forest";
 import coreRenderers from "@times-components/markup";
 import PullQuote from "@times-components/pull-quote";
-import { colours, spacing } from "@times-components/styleguide";
 import Video from "@times-components/video";
 import BodyParagraph from "./article-body-paragraph";
 import ArticleLink from "./article-link";
 import InsetCaption from "./inset-caption";
-
-const styles = StyleSheet.create({
-  ad: {
-    borderTopColor: colours.functional.keyline,
-    borderBottomColor: colours.functional.keyline,
-    borderTopWidth: 1,
-    borderBottomWidth: 1,
-    paddingTop: spacing(2),
-    paddingBottom: spacing(2),
-    marginBottom: spacing(4)
-  },
-  interactiveContainer: {
-    paddingRight: spacing(2),
-    paddingLeft: spacing(2),
-    marginBottom: spacing(4)
-  },
-  primaryContainer: {
-    width: "100%",
-    flexDirection: "column",
-    paddingBottom: spacing(5)
-  }
-});
+import styles from "../styles/article-body";
 
 const ArticleRow = ({
   content: { data, index },
